+++ example/lib/main-localized.dart
@@ -36,6 +36,7 @@ class Demo extends StatelessWidget {
         const Locale('hu', ''), // Hungarian, no country code
         const Locale('id', ''), // Indonesian, no country code
         const Locale('it', ''), // Italian, no country code
+        const Locale('kk', ''), // Kazakh, no country code
         const Locale('ko', ''), // Korean, no country code
         const Locale('pt', ''), // Portuguese, no country code
         const Locale('pl', ''), // Polish, no country code
@@ -120,6 +121,7 @@ class DemoLocalizationsDelegate
         'hu',
         'id',
         'it',
+        'kk',
         'ko',
         'pt',
         'pl',