+++ test/upgrader_test.dart
@@ -710,6 +710,7 @@ void main() {
     verifyMessages(UpgraderMessages(code: 'ru'), 'ru');
     verifyMessages(UpgraderMessages(code: 'ta'), 'ta');
     verifyMessages(UpgraderMessages(code: 'tr'), 'tr');
+    verifyMessages(UpgraderMessages(code: 'uk'), 'uk');
     verifyMessages(UpgraderMessages(code: 'vi'), 'vi');
   }, skip: false);
 }