+++ lib/src/upgrade_alert.dart
@@ -62,9 +62,11 @@ class UpgradeAlert extends UpgradeBase {
     }
 
     return FutureBuilder(
-        future: Upgrader().initialize(),
+        future: state.initialized,
         builder: (BuildContext context, AsyncSnapshot<bool> processed) {
-          if (processed.connectionState == ConnectionState.done) {
+          if (processed.connectionState == ConnectionState.done &&
+              processed.data != null &&
+              processed.data!) {
             Upgrader().checkVersion(context: context);
           }
           return child!;