+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+## 3.4.1
+
+* Fixed issue with Appcast where the upgrade message was not displayed.
+
 ## 3.4.0
 
 * Added new language translations for Tamil ('ta'), Kazakh ('kk'), Bengali ('bn'), Ukrainian ('uk').
 
 ## 3.3.1
 
-* Fixed the parsing of the appcast body to handle UTF-8 correctly.
+* Fixed the parsing of the Appcast body to handle UTF-8 correctly.
 
 ## 3.3.0